Move zodbot tests to their own file, and make the long_form part conditional like the others.
